# client-api-specs/core-business-service-internal-openapi-spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Internal API",
    "description": "IF-009 - Internal API",
    "version": "3.25.0-SNAPSHOT"
  },
  "servers": [
    {
      "url": "http://localhost:8200",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearer-jwt": []
    }
  ],
...
  "paths": {
...
    "/api/v1/internal/submissions/trust-add-dids-submissions/{trustAddDidsSubmissionId}": {
      "get": {
        "tags": [
          "TrustAddDidsSubmission Internal"
        ],
        "summary": "Get submission with proof of possession status",
        "operationId": "getSubmission",
        "parameters": [
          {
            "name": "trustAddDidsSubmissionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrustAdditionalDidsSubmissionInternalDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "errorCode": {
            "$ref": "#/components/schemas/ApiErrorCode"
          },
          "message": {
            "type": "string"
          },
          "additionalDetails": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ApiErrorCode": {
        "type": "string",
        "description": "       | Value                                       | Description                                                                           |\n       | ------------------------------------------- | ------------------------------------------------------------------------------------- |\n       | data_invalid                                | Indicates that the given data was invalid in syntax or semantic.                      |\n       | data_invalid_virus_detected                 | Indicates that the given data contains a virus.                                       |\n       | resource_not_found                          | Indicates that the requested data was not found in the given context.                 |\n       | resource_forbidden                          | Indicates that the requested data is not available with the permissions used.         |\n       | action_forbidden                            | Indicates that the requested action could not be processed with the permissions used. |\n       | partner_is_not_governmental                 | Indicates that the partner is not governmental.                                       |\n       | business_data_integrity_violation           | Indicates that a business data integrity violation occurred.                          |\n       | crypto_integrity_validation_failed          | Indicates that the crypto integrity validation failed.                                |\n       | did_resolve_failed                          | Indicates that the DID resolution failed.                                             |\n       | document_not_found                          | Indicates that the document was not found.                                            |\n       | identifier_validation_failed                | Indicates that the identifier validation failed.                                      |\n       | invalid_pagination                          | Indicates that the pagination parameters are invalid.                                 |\n       | max_size_exceeded                           | Indicates that the maximum size was exceeded.                                         |\n       | object_count_limit_reached                  | Indicates that the object count limit was reached.                                    |\n       | status_list_validation_failed               | Indicates that the status list validation failed.                                     |\n       | trust_onboarding_document_validation_failed | Indicates that the trust onboarding document validation failed.                       |\n       | vc_schema_submission_not_found              | Indicates that the VC schema submission was not found.                                |\n       | vc_type_metadata_validation_failed          | Indicates that the VC type metadata validation failed.                                |\n",
        "enum": [
          "data_invalid",
          "data_invalid_virus_detected",
          "resource_not_found",
          "resource_forbidden",
          "action_forbidden",
          "partner_is_not_governmental",
          "business_data_integrity_violation",
          "crypto_integrity_validation_failed",
          "did_resolve_failed",
          "document_not_found",
          "identifier_validation_failed",
          "invalid_pagination",
          "max_size_exceeded",
          "object_count_limit_reached",
          "status_list_validation_failed",
          "trust_onboarding_document_validation_failed",
          "vc_schema_submission_not_found",
          "vc_type_metadata_validation_failed",
          "vqps_publication_failed",
          "vqps_publication_wait_timeout"
        ]
      },
...
      "ProofOfPossession": {
        "type": "object",
        "properties": {
          "did": {
            "type": "string",
            "minLength": 1
          },
          "nonce": {
            "type": "string",
            "minLength": 1
          },
          "status": {
            "$ref": "#/components/schemas/ProofOfPossessionStatus"
          },
          "verifiedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "did",
          "nonce",
          "status"
        ]
      },
      "ProofOfPossessionStatus": {
        "type": "string",
        "enum": [
          "VALID",
          "NOT_SUPPLIED"
        ]
      },
...
      "TrustAdditionalDidsSubmissionInternalDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "UNSUBMITTED",
              "UNSUBMITTED_TIMEOUT",
              "SUBMITTED",
              "SUCCEEDED",
              "REJECTED"
            ]
          },
          "permissionDid": {
            "$ref": "#/components/schemas/ProofOfPossession"
          },
          "didsToAdd": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProofOfPossession"
            }
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "didsToAdd",
          "id",
          "permissionDid",
          "status"
        ]
      },
...
    },
    "securitySchemes": {
      "bearer-jwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "jwt"
      },
      "prometheus": {
        "type": "http",
        "scheme": "basic"
      },
      "OIDC": {
        "type": "openIdConnect",
        "description": "OAuth2-Authentication",
        "openIdConnectUrl": "https://swiyu-oauth-mock-server-d.apps.p-szb-ros-shrd-npr-01.cloud.admin.ch/.well-known/openid-configuration"
      }
    }
  }
}